% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS YEARs}
\usage{
fars_read_years(years)
}
\arguments{
\item{YEARs}{A vector with a list of YEARs}
}
\value{
A data.frame including entries in data by MONTH, or NULL if the
 \code{YEAR} is not valid
}
\description{
Ancillary function for \code{fars_summarize_years}
}
\examples{
\dontrun{
fars_read_years(2013)
}
}
\seealso{
\link{fars_read}

\link{make_filename}

\link{fars_summarize_years}
}
