% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data by YEARs}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{YEARs}{A vector with a list of YEARs to summarize by.}
}
\value{
A data.frame with number of accidents by YEARs summarized by MONTH
}
\description{
This function summarizes YEARly accidents data by MONTH
}
\examples{
\dontrun{
plot(fars_summarize_years(2015))
fars_summarize_years(c(2015, 2014))
}
}
\seealso{
\link{fars_read_years}
}
