% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Display accidents map by state and YEAR}
\usage{
fars_map_state(state_num, YEAR)
}
\arguments{
\item{state_num}{An Integer with the state Code}

\item{YEAR}{A string, or an integer, with the input \code{YEAR}}
}
\value{
None
}
\description{
Creates a plot with a state map including the accidents location by YEAR
If the \code{state_num} is invalid the function shows an error
}
\seealso{
\link{fars_read}

\link{make_filename}
}
